Moving to a more generic way to handle extraVol
Secrets, configMaps and other entities can be passed to the pods via API
extension.
The operator can import the shared structure and provide both volumes
and mounts according to the crd.
This patch is the first pass to extend what was done with cephSecrets
to the new general purpose strategy.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>