Pass storage.ServiceType to GetVolumes
This patch aligns the code with the lib-common change, where the
ServiceType is provided for each Cinder component, and it can be
passed to GetVolume functions to be aware about the src that has
requested an extraVol mount.
Then, according to the Propagation policy passed in the CRD, the
extraMount will be (or not be) propagated to the storage-operator
services.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>